refactor(js): cleanup and restructure the nodejs
