+++ wrappers/javascript/nodejs/src/error.ts
@@ -2,18 +2,16 @@ import type { IndyVdrErrorObject } from 'indy-vdr-shared'
 
 import { IndyVdrError } from 'indy-vdr-shared'
 
-import { nativeIndyVdr } from './lib'
-import { allocateStringBuffer } from './utils'
+import { allocateString } from './ffi'
+import { nativeIndyVdr } from './library'
 
 export const handleError = (code: number) => {
-  if (code == 0) return
+  if (code === 0) return
 
-  const nativeError = allocateStringBuffer()
+  const nativeError = allocateString()
   nativeIndyVdr.indy_vdr_get_current_error(nativeError)
 
-  // TODO
-  // eslint-disable-next-line @typescript-eslint/no-unsafe-assignment, @typescript-eslint/no-unsafe-argument
-  const indyVdrErrorObject: IndyVdrErrorObject = JSON.parse(nativeError.deref())
+  const indyVdrErrorObject = JSON.parse(nativeError.deref() as string) as IndyVdrErrorObject
 
   throw new IndyVdrError(indyVdrErrorObject)
 }
