+++ wrappers/javascript/nodejs/src/ffi/conversion.ts
@@ -1,0 +1,11 @@
+import { ByteBufferArray, ByteBuffer } from './structures'
+
+export const uint8ArrayToByteBuffer = (typedArray: Buffer) => {
+  const len = typedArray.length
+  const data = new ByteBufferArray(typedArray)
+
+  return ByteBuffer({
+    len,
+    data,
+  })
+}
