+++ wrappers/javascript/nodejs/src/library/index.ts
@@ -1,0 +1,2 @@
+export * from './NativeBindings'
+export * from './register'
