--- wrappers/javascript/nodejs/src/utils/index.ts
@@ -1,2 +1,0 @@
-export * from './ffiTools'
-export * from './serialize'
